requirements.txt
alembic==1.7.6
APIFlask==0.11.0
apispec==5.1.1
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
marshmallow==3.14.1
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
webargs==8.1.0
Werkzeug==2.0.3
